LPD-75317 buildRest
