+++ modules/apps/headless/headless-admin-site/headless-admin-site-api/src/main/java/com/liferay/headless/admin/site/dto/v1_0/PageSettings.java
@@ -73,51 +73,6 @@ public abstract class PageSettings implements Serializable {
 	}
 
 	@io.swagger.v3.oas.annotations.media.Schema(
-		description = "A list of custom meta tags this page has."
-	)
-	@Valid
-	public CustomMetaTag[] getCustomMetaTags() {
-		if (_customMetaTagsSupplier != null) {
-			customMetaTags = _customMetaTagsSupplier.get();
-
-			_customMetaTagsSupplier = null;
-		}
-
-		return customMetaTags;
-	}
-
-	public void setCustomMetaTags(CustomMetaTag[] customMetaTags) {
-		this.customMetaTags = customMetaTags;
-
-		_customMetaTagsSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setCustomMetaTags(
-		UnsafeSupplier<CustomMetaTag[], Exception>
-			customMetaTagsUnsafeSupplier) {
-
-		_customMetaTagsSupplier = () -> {
-			try {
-				return customMetaTagsUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField(description = "A list of custom meta tags this page has.")
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected CustomMetaTag[] customMetaTags;
-
-	@JsonIgnore
-	private Supplier<CustomMetaTag[]> _customMetaTagsSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema(
 		description = "A flag that indicates whether the page is hidden from navigation."
 	)
 	public Boolean getHiddenFromNavigation() {
@@ -209,51 +164,6 @@ public abstract class PageSettings implements Serializable {
 	@JsonIgnore
 	private Supplier<SitePageNavigationSettings> _navigationSettingsSupplier;
 
-	@io.swagger.v3.oas.annotations.media.Schema(
-		description = "The page's Open Graph settings."
-	)
-	@Valid
-	public OpenGraphSettings getOpenGraphSettings() {
-		if (_openGraphSettingsSupplier != null) {
-			openGraphSettings = _openGraphSettingsSupplier.get();
-
-			_openGraphSettingsSupplier = null;
-		}
-
-		return openGraphSettings;
-	}
-
-	public void setOpenGraphSettings(OpenGraphSettings openGraphSettings) {
-		this.openGraphSettings = openGraphSettings;
-
-		_openGraphSettingsSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setOpenGraphSettings(
-		UnsafeSupplier<OpenGraphSettings, Exception>
-			openGraphSettingsUnsafeSupplier) {
-
-		_openGraphSettingsSupplier = () -> {
-			try {
-				return openGraphSettingsUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField(description = "The page's Open Graph settings.")
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected OpenGraphSettings openGraphSettings;
-
-	@JsonIgnore
-	private Supplier<OpenGraphSettings> _openGraphSettingsSupplier;
-
 	@DecimalMax("2147483647")
 	@DecimalMin("0")
 	@io.swagger.v3.oas.annotations.media.Schema(
@@ -300,50 +210,6 @@ public abstract class PageSettings implements Serializable {
 
 	@JsonIgnore
 	private Supplier<Integer> _prioritySupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema(
-		description = "The page's SEO settings."
-	)
-	@Valid
-	public SEOSettings getSeoSettings() {
-		if (_seoSettingsSupplier != null) {
-			seoSettings = _seoSettingsSupplier.get();
-
-			_seoSettingsSupplier = null;
-		}
-
-		return seoSettings;
-	}
-
-	public void setSeoSettings(SEOSettings seoSettings) {
-		this.seoSettings = seoSettings;
-
-		_seoSettingsSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setSeoSettings(
-		UnsafeSupplier<SEOSettings, Exception> seoSettingsUnsafeSupplier) {
-
-		_seoSettingsSupplier = () -> {
-			try {
-				return seoSettingsUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField(description = "The page's SEO settings.")
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected SEOSettings seoSettings;
-
-	@JsonIgnore
-	private Supplier<SEOSettings> _seoSettingsSupplier;
 
 	@io.swagger.v3.oas.annotations.media.Schema
 	@JsonGetter("type")
@@ -424,28 +290,6 @@ public abstract class PageSettings implements Serializable {
 
 		sb.append("{");
 
-		CustomMetaTag[] customMetaTags = getCustomMetaTags();
-
-		if (customMetaTags != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"customMetaTags\": ");
-
-			sb.append("[");
-
-			for (int i = 0; i < customMetaTags.length; i++) {
-				sb.append(String.valueOf(customMetaTags[i]));
-
-				if ((i + 1) < customMetaTags.length) {
-					sb.append(", ");
-				}
-			}
-
-			sb.append("]");
-		}
-
 		Boolean hiddenFromNavigation = getHiddenFromNavigation();
 
 		if (hiddenFromNavigation != null) {
@@ -470,18 +314,6 @@ public abstract class PageSettings implements Serializable {
 			sb.append(String.valueOf(navigationSettings));
 		}
 
-		OpenGraphSettings openGraphSettings = getOpenGraphSettings();
-
-		if (openGraphSettings != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"openGraphSettings\": ");
-
-			sb.append(String.valueOf(openGraphSettings));
-		}
-
 		Integer priority = getPriority();
 
 		if (priority != null) {
@@ -492,18 +324,6 @@ public abstract class PageSettings implements Serializable {
 			sb.append("\"priority\": ");
 
 			sb.append(priority);
-		}
-
-		SEOSettings seoSettings = getSeoSettings();
-
-		if (seoSettings != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"seoSettings\": ");
-
-			sb.append(String.valueOf(seoSettings));
 		}
 
 		Type type = getType();
