+++ modules/apps/headless/headless-admin-site/headless-admin-site-client/src/main/java/com/liferay/headless/admin/site/client/dto/v1_0/WidgetPageSettings.java
@@ -26,6 +26,28 @@ public class WidgetPageSettings
 		return WidgetPageSettingsSerDes.toDTO(json);
 	}
 
+	public CustomMetaTag[] getCustomMetaTags() {
+		return customMetaTags;
+	}
+
+	public void setCustomMetaTags(CustomMetaTag[] customMetaTags) {
+		this.customMetaTags = customMetaTags;
+	}
+
+	public void setCustomMetaTags(
+		UnsafeSupplier<CustomMetaTag[], Exception>
+			customMetaTagsUnsafeSupplier) {
+
+		try {
+			customMetaTags = customMetaTagsUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected CustomMetaTag[] customMetaTags;
+
 	public Boolean getCustomizable() {
 		return customizable;
 	}
@@ -110,6 +132,49 @@ public class WidgetPageSettings
 	}
 
 	protected String layoutTemplateId;
+
+	public OpenGraphSettings getOpenGraphSettings() {
+		return openGraphSettings;
+	}
+
+	public void setOpenGraphSettings(OpenGraphSettings openGraphSettings) {
+		this.openGraphSettings = openGraphSettings;
+	}
+
+	public void setOpenGraphSettings(
+		UnsafeSupplier<OpenGraphSettings, Exception>
+			openGraphSettingsUnsafeSupplier) {
+
+		try {
+			openGraphSettings = openGraphSettingsUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected OpenGraphSettings openGraphSettings;
+
+	public SEOSettings getSeoSettings() {
+		return seoSettings;
+	}
+
+	public void setSeoSettings(SEOSettings seoSettings) {
+		this.seoSettings = seoSettings;
+	}
+
+	public void setSeoSettings(
+		UnsafeSupplier<SEOSettings, Exception> seoSettingsUnsafeSupplier) {
+
+		try {
+			seoSettings = seoSettingsUnsafeSupplier.get();
+		}
+		catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	protected SEOSettings seoSettings;
 
 	public ItemExternalReference getWidgetPageTemplateReference() {
 		return widgetPageTemplateReference;
