+++ modules/apps/headless/headless-admin-site/headless-admin-site-client/src/main/java/com/liferay/headless/admin/site/client/serdes/v1_0/ContentPageSettingsSerDes.java
@@ -70,6 +70,27 @@ public class ContentPageSettingsSerDes {
 			sb.append("]");
 		}
 
+		if (contentPageSettings.getOpenGraphSettings() != null) {
+			if (sb.length() > 1) {
+				sb.append(", ");
+			}
+
+			sb.append("\"openGraphSettings\": ");
+
+			sb.append(
+				String.valueOf(contentPageSettings.getOpenGraphSettings()));
+		}
+
+		if (contentPageSettings.getSeoSettings() != null) {
+			if (sb.length() > 1) {
+				sb.append(", ");
+			}
+
+			sb.append("\"seoSettings\": ");
+
+			sb.append(String.valueOf(contentPageSettings.getSeoSettings()));
+		}
+
 		if (contentPageSettings.getHiddenFromNavigation() != null) {
 			if (sb.length() > 1) {
 				sb.append(", ");
@@ -91,17 +112,6 @@ public class ContentPageSettingsSerDes {
 				String.valueOf(contentPageSettings.getNavigationSettings()));
 		}
 
-		if (contentPageSettings.getOpenGraphSettings() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"openGraphSettings\": ");
-
-			sb.append(
-				String.valueOf(contentPageSettings.getOpenGraphSettings()));
-		}
-
 		if (contentPageSettings.getPriority() != null) {
 			if (sb.length() > 1) {
 				sb.append(", ");
@@ -110,16 +120,6 @@ public class ContentPageSettingsSerDes {
 			sb.append("\"priority\": ");
 
 			sb.append(contentPageSettings.getPriority());
-		}
-
-		if (contentPageSettings.getSeoSettings() != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"seoSettings\": ");
-
-			sb.append(String.valueOf(contentPageSettings.getSeoSettings()));
 		}
 
 		if (contentPageSettings.getType() != null) {
@@ -164,6 +164,24 @@ public class ContentPageSettingsSerDes {
 				String.valueOf(contentPageSettings.getCustomMetaTags()));
 		}
 
+		if (contentPageSettings.getOpenGraphSettings() == null) {
+			map.put("openGraphSettings", null);
+		}
+		else {
+			map.put(
+				"openGraphSettings",
+				String.valueOf(contentPageSettings.getOpenGraphSettings()));
+		}
+
+		if (contentPageSettings.getSeoSettings() == null) {
+			map.put("seoSettings", null);
+		}
+		else {
+			map.put(
+				"seoSettings",
+				String.valueOf(contentPageSettings.getSeoSettings()));
+		}
+
 		if (contentPageSettings.getHiddenFromNavigation() == null) {
 			map.put("hiddenFromNavigation", null);
 		}
@@ -182,30 +200,12 @@ public class ContentPageSettingsSerDes {
 				String.valueOf(contentPageSettings.getNavigationSettings()));
 		}
 
-		if (contentPageSettings.getOpenGraphSettings() == null) {
-			map.put("openGraphSettings", null);
-		}
-		else {
-			map.put(
-				"openGraphSettings",
-				String.valueOf(contentPageSettings.getOpenGraphSettings()));
-		}
-
 		if (contentPageSettings.getPriority() == null) {
 			map.put("priority", null);
 		}
 		else {
 			map.put(
 				"priority", String.valueOf(contentPageSettings.getPriority()));
-		}
-
-		if (contentPageSettings.getSeoSettings() == null) {
-			map.put("seoSettings", null);
-		}
-		else {
-			map.put(
-				"seoSettings",
-				String.valueOf(contentPageSettings.getSeoSettings()));
 		}
 
 		if (contentPageSettings.getType() == null) {
@@ -236,6 +236,12 @@ public class ContentPageSettingsSerDes {
 			if (Objects.equals(jsonParserFieldName, "customMetaTags")) {
 				return false;
 			}
+			else if (Objects.equals(jsonParserFieldName, "openGraphSettings")) {
+				return false;
+			}
+			else if (Objects.equals(jsonParserFieldName, "seoSettings")) {
+				return false;
+			}
 			else if (Objects.equals(
 						jsonParserFieldName, "hiddenFromNavigation")) {
 
@@ -246,13 +252,7 @@ public class ContentPageSettingsSerDes {
 
 				return false;
 			}
-			else if (Objects.equals(jsonParserFieldName, "openGraphSettings")) {
-				return false;
-			}
 			else if (Objects.equals(jsonParserFieldName, "priority")) {
-				return false;
-			}
-			else if (Objects.equals(jsonParserFieldName, "seoSettings")) {
 				return false;
 			}
 			else if (Objects.equals(jsonParserFieldName, "type")) {
@@ -283,6 +283,19 @@ public class ContentPageSettingsSerDes {
 					contentPageSettings.setCustomMetaTags(customMetaTagsArray);
 				}
 			}
+			else if (Objects.equals(jsonParserFieldName, "openGraphSettings")) {
+				if (jsonParserFieldValue != null) {
+					contentPageSettings.setOpenGraphSettings(
+						OpenGraphSettingsSerDes.toDTO(
+							(String)jsonParserFieldValue));
+				}
+			}
+			else if (Objects.equals(jsonParserFieldName, "seoSettings")) {
+				if (jsonParserFieldValue != null) {
+					contentPageSettings.setSeoSettings(
+						SEOSettingsSerDes.toDTO((String)jsonParserFieldValue));
+				}
+			}
 			else if (Objects.equals(
 						jsonParserFieldName, "hiddenFromNavigation")) {
 
@@ -300,23 +313,10 @@ public class ContentPageSettingsSerDes {
 							(String)jsonParserFieldValue));
 				}
 			}
-			else if (Objects.equals(jsonParserFieldName, "openGraphSettings")) {
-				if (jsonParserFieldValue != null) {
-					contentPageSettings.setOpenGraphSettings(
-						OpenGraphSettingsSerDes.toDTO(
-							(String)jsonParserFieldValue));
-				}
-			}
 			else if (Objects.equals(jsonParserFieldName, "priority")) {
 				if (jsonParserFieldValue != null) {
 					contentPageSettings.setPriority(
 						Integer.valueOf((String)jsonParserFieldValue));
-				}
-			}
-			else if (Objects.equals(jsonParserFieldName, "seoSettings")) {
-				if (jsonParserFieldValue != null) {
-					contentPageSettings.setSeoSettings(
-						SEOSettingsSerDes.toDTO((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(jsonParserFieldName, "type")) {
