+++ modules/apps/headless/headless-admin-site/headless-admin-site-client/src/main/java/com/liferay/headless/admin/site/client/serdes/v1_0/PageSettingsSerDes.java
@@ -6,7 +6,6 @@
 package com.liferay.headless.admin.site.client.serdes.v1_0;
 
 import com.liferay.headless.admin.site.client.dto.v1_0.ContentPageSettings;
-import com.liferay.headless.admin.site.client.dto.v1_0.CustomMetaTag;
 import com.liferay.headless.admin.site.client.dto.v1_0.PageSetPageSettings;
 import com.liferay.headless.admin.site.client.dto.v1_0.PageSettings;
 import com.liferay.headless.admin.site.client.dto.v1_0.WidgetPageSettings;
@@ -87,15 +86,6 @@ public class PageSettingsSerDes {
 
 		Map<String, String> map = new TreeMap<>();
 
-		if (pageSettings.getCustomMetaTags() == null) {
-			map.put("customMetaTags", null);
-		}
-		else {
-			map.put(
-				"customMetaTags",
-				String.valueOf(pageSettings.getCustomMetaTags()));
-		}
-
 		if (pageSettings.getHiddenFromNavigation() == null) {
 			map.put("hiddenFromNavigation", null);
 		}
@@ -114,28 +104,11 @@ public class PageSettingsSerDes {
 				String.valueOf(pageSettings.getNavigationSettings()));
 		}
 
-		if (pageSettings.getOpenGraphSettings() == null) {
-			map.put("openGraphSettings", null);
-		}
-		else {
-			map.put(
-				"openGraphSettings",
-				String.valueOf(pageSettings.getOpenGraphSettings()));
-		}
-
 		if (pageSettings.getPriority() == null) {
 			map.put("priority", null);
 		}
 		else {
 			map.put("priority", String.valueOf(pageSettings.getPriority()));
-		}
-
-		if (pageSettings.getSeoSettings() == null) {
-			map.put("seoSettings", null);
-		}
-		else {
-			map.put(
-				"seoSettings", String.valueOf(pageSettings.getSeoSettings()));
 		}
 
 		if (pageSettings.getType() == null) {
@@ -163,12 +136,7 @@ public class PageSettingsSerDes {
 
 		@Override
 		protected boolean parseMaps(String jsonParserFieldName) {
-			if (Objects.equals(jsonParserFieldName, "customMetaTags")) {
-				return false;
-			}
-			else if (Objects.equals(
-						jsonParserFieldName, "hiddenFromNavigation")) {
-
+			if (Objects.equals(jsonParserFieldName, "hiddenFromNavigation")) {
 				return false;
 			}
 			else if (Objects.equals(
@@ -176,13 +144,7 @@ public class PageSettingsSerDes {
 
 				return false;
 			}
-			else if (Objects.equals(jsonParserFieldName, "openGraphSettings")) {
-				return false;
-			}
 			else if (Objects.equals(jsonParserFieldName, "priority")) {
-				return false;
-			}
-			else if (Objects.equals(jsonParserFieldName, "seoSettings")) {
 				return false;
 			}
 			else if (Objects.equals(jsonParserFieldName, "type")) {
@@ -226,25 +188,7 @@ public class PageSettingsSerDes {
 			PageSettings pageSettings, String jsonParserFieldName,
 			Object jsonParserFieldValue) {
 
-			if (Objects.equals(jsonParserFieldName, "customMetaTags")) {
-				if (jsonParserFieldValue != null) {
-					Object[] jsonParserFieldValues =
-						(Object[])jsonParserFieldValue;
-
-					CustomMetaTag[] customMetaTagsArray =
-						new CustomMetaTag[jsonParserFieldValues.length];
-
-					for (int i = 0; i < customMetaTagsArray.length; i++) {
-						customMetaTagsArray[i] = CustomMetaTagSerDes.toDTO(
-							(String)jsonParserFieldValues[i]);
-					}
-
-					pageSettings.setCustomMetaTags(customMetaTagsArray);
-				}
-			}
-			else if (Objects.equals(
-						jsonParserFieldName, "hiddenFromNavigation")) {
-
+			if (Objects.equals(jsonParserFieldName, "hiddenFromNavigation")) {
 				if (jsonParserFieldValue != null) {
 					pageSettings.setHiddenFromNavigation(
 						(Boolean)jsonParserFieldValue);
@@ -259,23 +203,10 @@ public class PageSettingsSerDes {
 							(String)jsonParserFieldValue));
 				}
 			}
-			else if (Objects.equals(jsonParserFieldName, "openGraphSettings")) {
-				if (jsonParserFieldValue != null) {
-					pageSettings.setOpenGraphSettings(
-						OpenGraphSettingsSerDes.toDTO(
-							(String)jsonParserFieldValue));
-				}
-			}
 			else if (Objects.equals(jsonParserFieldName, "priority")) {
 				if (jsonParserFieldValue != null) {
 					pageSettings.setPriority(
 						Integer.valueOf((String)jsonParserFieldValue));
-				}
-			}
-			else if (Objects.equals(jsonParserFieldName, "seoSettings")) {
-				if (jsonParserFieldValue != null) {
-					pageSettings.setSeoSettings(
-						SEOSettingsSerDes.toDTO((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(jsonParserFieldName, "type")) {
